refactor(pyo3): move Python bindings to src/python.rs and register from lib.rs
